pyproject.toml
[tool.poetry]
name = "itmo"
version = "0.1.0"
description = ""
authors = ["sobolevn <mail@sobolevn.me>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
requests = "^2.22"
django = "^3.0.1"
djangorestframework = "^3.11.0"
hypothesis = "^4.56.1"

[tool.poetry.dev-dependencies]
wemake-python-styleguide = "^0.12.5"
flake8-pytest-style = "^0.1.3"
flake8-pytest = "^1.3"

safety = "^1.8"

pytest = "^5.2"
pytest-randomly = "^3.1"
pytest-repeat = "^0.8.0"
pytest-timeout = "^1.3"
mypy = "^0.740"
pytest-remotedata = "^0.3.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
